+++ src/ft_getenv.c
@@ -6,7 +6,7 @@
 /*   By: chlee2 <chlee2@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/10 14:27:48 by chlee2            #+#    #+#             */
-/*   Updated: 2025/03/30 21:05:12 by chlee2           ###   ########.fr       */
+/*   Updated: 2025/03/31 18:11:55 by chlee2           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,29 +54,29 @@ char	*ft_getenv_value_ptr(char *env_name, t_shell *shell)
 	return (0);
 }
 
-char	*ft_getenv(char *env_name, t_shell *shell)
-{
-	t_key_value	*pair;
-	t_key_value	*head;
-	char		*result;
+// char	*ft_getenv(char *env_name, t_shell *shell)
+// {
+// 	t_key_value	*pair;
+// 	t_key_value	*head;
+// 	char		*result;
 
-	result = NULL;
-	head = NULL;
-	pair = arr_to_key_value(shell);
-	head = pair;
-	while (head)
-	{
-		if (ft_strcmp(head->key, env_name) == 0)
-		{
-			result = ft_strdup(head->value);
-			if (!result)
-			//TODO
+// 	result = NULL;
+// 	head = NULL;
+// 	pair = arr_to_key_value(shell);
+// 	head = pair;
+// 	while (head)
+// 	{
+// 		if (ft_strcmp(head->key, env_name) == 0)
+// 		{
+// 			result = ft_strdup(head->value);
+// 			if (!result)
+// 			//TODO
 			
-				malloc_fail_clean_exit(shell);
-			break ;
-		}
-		head = head->next;
-	}
-	free_key_value_list(pair);
-	return (result);
-}
+// 				malloc_fail_clean_exit(shell);
+// 			break ;
+// 		}
+// 		head = head->next;
+// 	}
+// 	free_key_value_list(pair);
+// 	return (result);
+// }
