+++ src/utils/lvl_handler.c
@@ -6,7 +6,7 @@
 /*   By: chlee2 <chlee2@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/27 11:45:26 by chlee2            #+#    #+#             */
-/*   Updated: 2025/03/30 21:48:22 by mbutuzov         ###   ########.fr       */
+/*   Updated: 2025/03/31 18:12:40 by chlee2           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 //new_entry == ne
 static int	replace_env_var(t_shell *shell, char **envp, char *key, char *ne)
 {
-	(void)envp;
-	int index;
+	int	index;
 
+	(void)envp;
 	index = get_envp_var_index(key, shell);
 	return (adjust_envp(ne, shell, index));
 }
